% Generated by roxygen2: do not edit by hand
% Please edit documentation in helpers/growth_rate.R
\name{growth_rate}
\alias{growth_rate}
\title{Helper function: compute growth rate of a time-series}
\usage{
growth_rate(
  ts,
  ...,
  delta = 1,
  type = c("geometric", "log", "difference"),
  units = c("percent", "decimal"),
  future_growth = c(FALSE, TRUE)
)
}
\description{
Helper function: compute growth rate of a time-series
}
