% Generated by roxygen2: do not edit by hand
% Please edit documentation in helpers/growth_rate.R
\name{growth_rate_df}
\alias{growth_rate_df}
\title{Helper function: tidy compute growth rates over many periods of a subset of variables of a data frame}
\usage{
growth_rate_df(data, var_list, delta, ...)
}
\description{
Helper function: tidy compute growth rates over many periods of a subset of variables of a data frame
}
