% Generated by roxygen2: do not edit by hand
% Please edit documentation in helpers/hetreg.R
\name{hetreg}
\alias{hetreg}
\title{Heteroskedastic linear regression}
\usage{
hetreg(data, y, x, het = x, ..., gls_opt = NULL)
}
\arguments{
\item{data}{data frame containing the variables in `y`, `x` and `het`.}

\item{y}{string (character vector) with the name of the independent variable.}

\item{x}{vector of strings (character vectors) with the name of the
independent variable(s).}

\item{het}{vector of strings (character vectors) with the name of the
variable(s) that enter(s) the variance function as independent variable(s).}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. **Cautionary note:** Misspelled arguments will be
absorbed in `...`, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass `conf.lvel = 0.9`, all computation will
proceed using `conf.level = 0.95`.}

\item{gls_opt}{optional arguments passed to [`gls()`][nlme::gls()].}
}
\value{
an object of class `gls` representing the linear model fit.
See [`gls()`][nlme::gls()] for details.
}
\description{
`hetreg()` fits a heteroskedastic linear regression
where the variance is an exponential function of specified variables
}
\details{
`hetreg()` is a wrapper for [nlme::gls()]. Compared
to [`gls()`][nlme::gls()], the function `hetreg()`
* is pipeable (plays nice with the tidyverse)
* automatically constructs the [`gls()`][nlme::gls()] argument `model` from
`y` and `x`, and the [`gls()`][nlme::gls()] argument `weights` from `het`
* only allows for an exponential variance function structure defined by
a combination [`varComb()`][nlme::varComb()] of [varExp()][nlme::varExp()]
 terms
* has [broom][broom::broom()] tidiers

The model estimated is:
\deqn{y_i = x_i \beta + \epsilon_i \\
      var(\epsilon_i) = \exp(2 \theta v_i )}
where \eqn{y_i}, \eqn{i = 1,\dotsc,n}, is the dependent variable;
 \eqn{x_i = (x_{1i},x_{2i},\dotsc,x_{ki})} are the \eqn{k} independent
 variables that model the mean function;
  \eqn{v_i = (v_{1i},v_{2i},\dotsc,v_{mi})}
  are the \eqn{m} variables that model the variance function (corresponding to the
  argument `het`); \eqn{\beta} is a column vector of unknown parameters in
   the mean function, and \eqn{\theta} is a column vector of unknown
   parameters in the variance function.
 `hetreg()` is similar to the command
  [hetregress](https://www.stata.com/manuals/rhetregress.pdf)
 in Stata, where
 ```Stata
 hetregress x y, het(z)
 ```
 corresponds to the model
 \deqn{y_i = x_i \beta + \epsilon_i \\
      var(\epsilon_i) = \exp(z_i \alpha )}
The \eqn{z_i} map to \eqn{v_i} and \eqn{\alpha} maps to \eqn{2 \theta}.
}
