% Generated by roxygen2: do not edit by hand
% Please edit documentation in helpers/hetreg_tidiers.R
\name{tidy.het}
\alias{tidy.het}
\alias{het_tidiers}
\alias{glance.het}
\alias{augment.het}
\title{Tidying methods for heteroskedastic linear regressions}
\usage{
tidy.het(x, conf_int = FALSE, conf_level = 0.95, ...)

glance.het(x, ...)

augment.het(x, ...)
}
\arguments{
\item{x}{An object of class `het` created by [vfci::hetreg()].}

\item{conf_int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to `FALSE`.}

\item{conf_level}{The confidence level to use for the confidence interval
if `conf_int = TRUE`. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. **Cautionary note:** Misspelled arguments will be
absorbed in `...`, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass `conf.lvel = 0.9`, all computation will
proceed using `conf.level = 0.95`.}
}
\value{
Various statistics and post-estimation values for [hetreg()]
}
\description{
Tidying methods `tidy`, `glance` and `augment` for objects of class `het`
that are compatible with [broom][broom::broom()].
}
\section{Functions}{
\itemize{
\item \code{tidy.het()}: returns the same information as `nlme_tidiers()` from
[`broom.mixed`](https://github.com/bbolker/broom.mixed)
for the mean equation, and adds coefficients, standard errors, t-statistics
and p-values for the variance equation

\item \code{glance.het()}: same as
[`broom.mixed::glance.gls()`](https://github.com/bbolker/broom.mixed)

\item \code{augment.het()}: returns the standard deviation of the error term

}}
\seealso{
[broom::tidy()], `nlme_tidiers()` from
 [`broom.mixed`](https://github.com/bbolker/broom.mixed)
}
